Move weather-tool under weather and update dependencies
