+++ mcp-server/src/main/java/com/cyster/mcp/McpServer.java
@@ -10,7 +10,11 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.Bean;
 
 @SpringBootApplication(
-    scanBasePackageClasses = {McpServer.class, WeatherServicePackageMarker.class, WeatherToolPackageMarker.class})
+    scanBasePackageClasses = {
+      McpServer.class,
+      WeatherServicePackageMarker.class,
+      WeatherToolPackageMarker.class
+    })
 public class McpServer {
   public static void main(String[] args) {
     SpringApplication.run(McpServer.class, args);
