+++ settings.gradle
@@ -24,5 +24,5 @@ include "chat-flux-server"
 include "mcp-tool"
 include "chat-mvp-server"
 include "weather"
-include "weather-tool"
+include "weather:weather-tool"
