+++ weather/src/main/java/com/cyster/weather/impl/WeatherServiceImpl.java
@@ -4,7 +4,6 @@ import com.cyster.weather.service.WeatherService;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import java.util.List;
 import java.util.stream.Collectors;
-import org.springframework.ai.tool.annotation.Tool;
 import org.springframework.stereotype.Service;
 import org.springframework.web.client.RestClient;
