+++ weather/weather-tool/src/main/java/com/cyster/weather/tool/WeatherToolPackageMarker.java
@@ -2,5 +2,5 @@ package com.cyster.weather.tool;
 
 /** Marker class used to identify the base package for the weather-tool module. */
 public final class WeatherToolPackageMarker {
-    private WeatherToolPackageMarker() {}
+  private WeatherToolPackageMarker() {}
 }
